newest code update

+++ photo_system/system_master.sh
@@ -3,7 +3,10 @@ trap "trap - SIGTERM && kill -- -$$" SIGINT SIGTERM EXIT
 # This script starts all other relevant scripts for this code
 
 # Reset Queue
-rm Queue.txt
+
+if test -d Queue.txt; then
+  rm Queue.txt
+fi
 
 #  This runs the motion detector
 ./motion_detect.sh&
@@ -18,5 +21,5 @@ echo "time trigger initiated"
 echo "wildlife trigger initiated"
 ./wildlife_trigger.sh&
 
-
+echo ""
 wait
